fixes for compilation on 1.4
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@27 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ sccp_softkeys.c
@@ -287,7 +287,7 @@ void sccp_sk_gpickup(sccp_device_t * d, sccp_line_t * l, sccp_channel_t * c) {
 
 	sccp_log(10)(VERBOSE_PREFIX_3 "%s: Pickup the call from %s\n", d->id, original->name);
 
-	res = (ast->blocker);
+	res = (int) (ast->blocker);
 	
 	ast_mutex_unlock(&ast->lock);
